feat: isolate credit card payment method
